+++ torchbearer/__init__.py
@@ -23,7 +23,7 @@ Utilities
 
 from .version import __version__
 from . import magics
-from .bases import *
+from .bases import no_grad, enable_grad, cite, base_closure, Callback, Metric
 from .state import *
 
 from . import metrics